+++ src/contexts/GlobalContext.js
@@ -1,0 +1,9 @@
+
+import {createContext} from "react";
+
+export const GlobalContext = createContext()
+
+
+
+
+
