+++ src/pages/HomePage/HomePage.js
@@ -2,8 +2,13 @@ import Card from "../../components/Card/Card";
 import { Container } from "./HomePage.styled";
 import Header from "../../components/Header/Header";
 
-function HomePage(props) {
-  const { pokelist, addToPokedex, pokedex } = props;
+import {useContext} from "react";
+import {GlobalContext} from "../../contexts/GlobalContext"
+
+function HomePage() {
+
+  const context = useContext(GlobalContext)
+  const  {pokelist, addToPokedex,pokedex} = context
 
   // não mostrar pokemons que estão na pokedex
   const filteredPokelist = () =>
