+++ src/pages/PokedexPage/PokedexPage.js
@@ -3,8 +3,14 @@ import { Container } from "./PokedexPage.styled";
 import Header from "../../components/Header/Header";
 import { BASE_URL } from "../../constants/url";
 
-function PokedexPage(props) {
-  const { pokedex, removeFromPokedex } = props;
+import{useContext} from "react"
+import { GlobalContext } from "../../contexts/GlobalContext";
+
+function PokedexPage() {
+
+  const context = useContext (GlobalContext)
+
+  const { pokedex, removeFromPokedex } = context;
 
   return (
     <Container>
